requirements.txt
aiogram
apscheduler
SQLAlchemy
psycopg2-binary   # для PostgreSQL, если выберете PostgreSQL
openpyxl          # для импорта данных из Excel
google-cloud-texttospeech
python-dotenv     # для работы с .env файлами (если понадобится)
